requirements.txt
python-telegram-bot==13.15
opencv-python
python-dotenv
requests
numpy
